+++ NEWS.md
@@ -109,8 +109,9 @@ This page describes the most important changes in `LazyAlgebra`. The format is b
 - `FlexibleMatrix` is a pseudo-matrix whose number of row dimensions depends on its
   input argument.
 
-- Exported function `with_precision(T, A)` to change the numerical precision of `A` to be
-  that of the floating-point type `T`.
+- Exported functions `get_precision(A)` and `with_precision(T, A)` to retrieve the
+  numerical precision of `A` and to change the numerical precision of `A` to be the
+  floating-point type `T`.
 
 - New methods `vones!`, `vnans`, and `vnans!` to fill an array with ones or NaNs.